#3 Changes made to loading images.
